requirements.txt
# Core Framework
crewai==0.28.8
crewai-tools==0.1.6
langchain-openai==0.0.5

# Streamlit UI
streamlit==1.31.0
streamlit-chat==0.1.1

# Medical Image Processing
pydicom==2.4.4
nibabel==5.2.0
opencv-python==4.9.0.80
Pillow==10.2.0

# Document Processing
pdfplumber==0.10.3
pytesseract==0.3.10
fpdf==1.7.2

# PubMed Integration
biopython==1.83

# Data Processing
numpy==1.26.3
pandas==2.2.0
matplotlib==3.8.2

# Environment Management
python-dotenv==1.0.1

# Additional utilities
pathlib

transformer